+++ core/src/display_object/container.rs
@@ -493,7 +493,10 @@ pub trait TDisplayObjectContainer<'gc>:
                 // Non-visible objects and their children are excluded from tab ordering.
                 continue;
             }
-            if child.is_tabbable(context) {
+            if child
+                .as_interactive()
+                .is_some_and(|o| o.is_tabbable(context))
+            {
                 tab_order.push(child);
             }
             if let Some(container) = child.as_container() {
